Changed the screen boundaries.
Now the player can drift out in space (lose) instead of just being stoped at the edge of the screen. But I added boundaries to the zoom camera, when the camera is zoomed in the player is stoped at the edge of the main camera view instead of letting the lander drift out in space.